% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{naglr_pal}
\alias{naglr_pal}
\alias{scale_color_naglr}
\title{A qualitative color palette}
\usage{
naglr_pal()

scale_color_naglr(...)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}
}
\description{
A qualitative color palette

Discrete color & fill scales
}
\examples{
library(scales)
scales::show_col(naglr_pal()(8))
}
