% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme.R
\name{theme_naglr}
\alias{theme_naglr}
\title{A [ggplot2] theme based on hrbrthemes}
\usage{
theme_naglr(base_family = "Roboto Condensed", base_size = 11,
  plot_title_family = base_family, plot_title_size = 18,
  plot_title_face = "bold", plot_title_margin = 10,
  subtitle_family = "Roboto Condensed Light", subtitle_size = 12,
  subtitle_face = "plain", subtitle_margin = 15,
  strip_text_family = base_family, strip_text_size = 12,
  strip_text_face = "plain", caption_family = "Roboto Condensed Light",
  caption_size = 9, caption_face = "plain", caption_margin = 10,
  axis_title_family = base_family, axis_title_size = 9,
  axis_title_face = "plain", axis_title_just = "rt",
  plot_margin = margin(30, 30, 30, 30), grid = TRUE, axis = FALSE,
  ticks = FALSE)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{plot_title_family, plot_title_face, plot_title_size, plot_title_margin}{plot tilte family, face, size and margi}

\item{subtitle_family, subtitle_face, subtitle_size}{plot subtitle family, face
and size}

\item{subtitle_margin}{plot subtitle margin bottom (single numeric value)}

\item{strip_text_family, strip_text_face, strip_text_size}{facet label font
family, face and size}

\item{caption_family, caption_face, caption_size, caption_margin}{plot caption
family, face, size and margin}

\item{axis_title_family, axis_title_face, axis_title_size}{axis title font
family, face and size}

\item{axis_title_just}{axis title font justificationk one of `[blmcrt]`}

\item{plot_margin}{plot margin (specify with [ggplot2::margin])}

\item{grid}{panel grid (`TRUE`, `FALSE`, or a combination of `X`, `x`, `Y`,
`y`)}

\item{axis}{add x or y axes? `TRUE`, `FALSE`, "`xy`"}

\item{ticks}{ticks if `TRUE` add ticks}
}
\description{
Most of this theme and documentation is a simple copy & paste from
\url{https://github.com/hrbrmstr/hrbrthemes} with only few modifications. The
panel grid has other defaults (thicker, but greyer) so that plots look still
nice when using the [Cairo] device.
}
\details{
You should [import_roboto_condensed]() first and also install the fonts on
your system before trying to use this theme.
}
\examples{
\dontrun{
library(ggplot2)
library(dplyr)

# seminal scatterplot
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  labs(x="Fuel effiiency (mpg)", y="Weight (tons)",
       title="Seminal ggplot2 scatterplot example",
       subtitle="A plot that is only useful for demonstration purposes",
       caption="Brought to you by the letter 'g'") +
  theme_naglr()

# seminal bar chart

update_geom_font_defaults(family=font_rc_light)

count(mpg, class) \%>\%
  ggplot(aes(class, n)) +
  geom_col() +
  geom_text(aes(label=n), nudge_y=3) +
  labs(x="Fuel effiiency (mpg)", y="Weight (tons)",
       title="Seminal ggplot2 bar chart example",
       subtitle="A plot that is only useful for demonstration purposes",
       caption="Brought to you by the letter 'g'") +
  theme_naglr(grid="Y") +
  theme(axis.text.y=element_blank())
}

}
