% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{tint}
\alias{tint}
\alias{shade}
\title{Make colors brighter or darker}
\usage{
tint(x, fac)

shade(x, fac)
}
\arguments{
\item{x}{a color}

\item{fac}{the factor (between 0 and 1).}
}
\description{
Make colors brighter or darker
}
